refactor: change LoginScreen file name
